break up dirsnapshot.py
