+++ codex/librarian/watchdog/emitter.py
@@ -17,10 +17,8 @@ from watchdog.events import (
 from watchdog.observers.api import EventEmitter
 from watchdog.utils.dirsnapshot import DirectorySnapshot
 
-from codex.librarian.watchdog.dirsnapshot import (
-    CodexDatabaseSnapshot,
-    CodexDirectorySnapshotDiff,
-)
+from codex.librarian.watchdog.db_snapshot import CodexDatabaseSnapshot
+from codex.librarian.watchdog.dir_snapshot_diff import CodexDirectorySnapshotDiff
 from codex.librarian.watchdog.status import WatchdogStatusTypes
 from codex.models import Library
 from codex.status import Status
